Backend/python/src/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==2.1.1
python-dotenv==1.0.0
python-multipart
python-jose[cryptography]
passlib
pymongo